move login into shared components and remove deprecated stuff
