--- src/components/login/login-image-slider.tsx
@@ -1,56 +1,0 @@
-// node_modules
-import { FC, useEffect, useState } from "react";
-import styles from "./login-image-slider.module.scss";
-
-const CollectInformationSVG = () => {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" data-name="Layer 1" width="655.04715" height="473.4057" viewBox="0 0 655.04715 473.4057" xmlnsXlink="http://www.w3.org/1999/xlink"><path d="M787.31056,255.89377c-1.65466-12.13655-3.72987-25.16694-12.28987-33.92808a28.78566,28.78566,0,0,0-41.59054.43047c-6.23108,6.64735-8.98285,15.98128-9.23469,25.089s1.77661,18.11434,4.02681,26.94334a83.04327,83.04327,0,0,0,34.65826-6.82181,55.84219,55.84219,0,0,1,8.30984-3.26239c2.88751-.73493,4.94581,1.11445,7.63182,2.404l1.47445-2.82319c1.207,2.24618,4.94435,1.58037,6.3-.57936C787.95205,261.18611,787.65511,258.42024,787.31056,255.89377Z" transform="translate(-272.47643 -213.29715)" fill="#2f2e41" /><path d="M737.57332,307.13788l3.44812-17.61353L715.1681,278.3319l-46.21928-22.143a10.27075,10.27075,0,1,0-7.44548,11.5L705.511,294.84741Z" transform="translate(-272.47643 -213.29715)" fill="#ffb8b8" /><path d="M720.21831,304.15781l18.725-19.1418.37023.11261c.23288.07053,23.49781,7.16307,38.32927,14.37044a12.63351,12.63351,0,0,1,6.96921,8.03865,13.2503,13.2503,0,0,1-4.15554,13.92014,14.17281,14.17281,0,0,1-15.15882,2.41334Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><polygon points="447.684 175.017 500.338 189.087 510.971 139.346 462.642 124.387 447.684 175.017" fill="#ffb8b8" /><path d="M778.708,374.81334l-51.4362-11.82883L717.42211,339.112a18.26275,18.26275,0,0,1,6.58613-18.61415l18.79293-15.86613,10.07455-12.86951-1.09686-6.82015a5.9078,5.9078,0,0,1,3.93786-6.4325l13.42683-4.815a6.2481,6.2481,0,0,1,7.88115,3.27734l1.42276,3.29287s6.07151,2.85713,9.4948,4.46428a13.74125,13.74125,0,0,1,7.43067,8.41824,69.65979,69.65979,0,0,1-3.9496,51.1518Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><path d="M321.26739,628.22726l-24.86987-6.655a83.06563,83.06563,0,0,1-1.23506-39.94776c10.85,21.44784,39.98979,25.4917,57.25227,42.21709A49.96443,49.96443,0,0,1,367.073,664.59315l5.95764,17.2599a83.72444,83.72444,0,0,1-62.45381-31.84171,80.8743,80.8743,0,0,1-10.28587-17.488C310.69751,630.99881,321.26739,628.22726,321.26739,628.22726Z" transform="translate(-272.47643 -213.29715)" fill="#f2f2f2" /><path d="M630.24665,686.70285H386.48931a24.668,24.668,0,0,1-24.63995-24.64012V339.30027a24.668,24.668,0,0,1,24.64-24.64012H630.24665a24.668,24.668,0,0,1,24.64013,24.64012V662.06273A24.668,24.668,0,0,1,630.24665,686.70285Z" transform="translate(-272.47643 -213.29715)" fill="#e6e6e6" /><path d="M537.21155,666.72452h-133.588a22.46122,22.46122,0,0,1-22.43542-22.436V357.0741a22.46075,22.46075,0,0,1,22.43542-22.43541h209.4887a22.46075,22.46075,0,0,1,22.43542,22.43541V568.38842A98.44763,98.44763,0,0,1,537.21155,666.72452Z" transform="translate(-272.47643 -213.29715)" fill="#fff" /><path d="M592.21441,389.77863H424.52152a5.469,5.469,0,1,1-.01368-10.938H592.21441a5.469,5.469,0,0,1,.01368,10.938Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><path d="M592.21441,411.38946H424.52152a5.46929,5.46929,0,0,1,0-10.93858H592.21441a5.46929,5.46929,0,0,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><path d="M497.21233,433.00029H424.52149a5.46929,5.46929,0,1,1,0-10.93858h72.69084a5.46929,5.46929,0,0,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><path d="M592.21441,478.26142H424.52152a5.469,5.469,0,1,1-.01368-10.938H592.21441a5.469,5.469,0,0,1,.01368,10.938Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><path d="M592.21441,499.87225H424.52152a5.46929,5.46929,0,0,1,0-10.93858H592.21441a5.46929,5.46929,0,0,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><path d="M497.21233,521.483H424.52149a5.46929,5.46929,0,1,1,0-10.93858h72.69084a5.46929,5.46929,0,1,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><path d="M592.21441,566.74414H424.52152a5.469,5.469,0,1,1-.01368-10.938H592.21441a5.469,5.469,0,0,1,.01368,10.938Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><path d="M592.21441,588.355H424.52152a5.46929,5.46929,0,0,1,0-10.93858H592.21441a5.46929,5.46929,0,0,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><path d="M497.21233,609.9658H424.52149a5.4693,5.4693,0,0,1,0-10.93858h72.69084a5.4693,5.4693,0,0,1,0,10.93858Z" transform="translate(-272.47643 -213.29715)" fill="#e4e4e4" /><circle cx="370.37338" cy="107.64026" r="71.09969" fill="#007AFF" /><path d="M642.84977,382.18768A71.09,71.09,0,0,1,571.945,316.09274c-.11525,1.65523-.19489,3.32029-.19489,5.00488a71.09968,71.09968,0,0,0,142.19935,0c0-1.68459-.07964-3.34965-.19489-5.00488A71.09005,71.09005,0,0,1,642.84977,382.18768Z" transform="translate(-272.47643 -213.29715)" fill="#231f20" opacity="0.16" /><path d="M642.85,355.6823a8.91461,8.91461,0,0,1-8.9043-8.9043V295.257a8.90406,8.90406,0,1,1,17.80811,0v51.521A8.9142,8.9142,0,0,1,642.85,355.6823Z" transform="translate(-272.47643 -213.29715)" fill="#fff" /><path d="M608.185,321.01775a8.91461,8.91461,0,0,1,8.90429-8.9043h51.521a8.90406,8.90406,0,1,1,0,17.80811h-51.521A8.91419,8.91419,0,0,1,608.185,321.01775Z" transform="translate(-272.47643 -213.29715)" fill="#fff" /><path d="M749.73026,336.12525l-15.12956-9.65517-19.87785,19.9634-37.52231,34.90863a10.27075,10.27075,0,1,0,7.97721,11.13774L726.55817,361.465Z" transform="translate(-272.47643 -213.29715)" fill="#ffb8b8" /><path d="M740.60549,351.18538,729.6448,326.75426l.24-.30343c.151-.19052,15.2663-19.24629,27.40134-30.41072a12.63381,12.63381,0,0,1,10.03145-3.54358,13.25745,13.25745,0,0,1,11.43348,8.9617,14.1712,14.1712,0,0,1-3.30149,14.98926Z" transform="translate(-272.47643 -213.29715)" fill="#007AFF" /><circle cx="478.03159" cy="33.80024" r="23.78567" fill="#ffb8b8" /><polygon points="441.077 456.369 431.244 456.369 426.564 418.442 441.077 418.442 441.077 456.369" fill="#ffb8b8" /><path d="M718.74308,686.70285h-7.05162l-1.25869-6.65757-3.22367,6.65757H688.50652a4.20417,4.20417,0,0,1-2.389-7.66381l14.93533-10.31487V661.9936l15.70936.93763Z" transform="translate(-272.47643 -213.29715)" fill="#2f2e41" /><polygon points="564.317 448.015 555.181 451.652 536.806 418.145 550.29 412.777 564.317 448.015" fill="#ffb8b8" /><path d="M847.91639,675.22176l-6.55162,2.608-3.63169-5.72-.53284,7.37776-17.37645,6.917a4.20417,4.20417,0,0,1-5.054-6.23686l10.06143-15.10722-2.48926-6.25334,14.94226-4.93885Z" transform="translate(-272.47643 -213.29715)" fill="#2f2e41" /><path d="M717.34637,652.24291H697.00672l5.62123-95.55913-4.10943-64.77159.00864-.05188L713.85358,399.013a39.21218,39.21218,0,0,1,9.39572-19.69433l.18658-.21025,51.29147,12.66779,5.15705,11.759c17.60912,10.47337,23.17029,69.53271,23.39874,72.04843l7.45523,73.62394,18.23624,90.2337-12.84255,6.91459-.2148-.42778-40.54277-80.61954-25.6041-85.0184-16.39677,76.51829Z" transform="translate(-272.47643 -213.29715)" fill="#2f2e41" /><path d="M779.96517,248.21676c-.84171-5.52877-1.7208-11.10385-3.69814-16.30974s-5.16459-10.082-9.73665-13.03107c-7.23909-4.66912-16.61147-3.76861-24.72769-1.08281-6.27661,2.07709-12.40368,5.23408-16.82988,10.30946-4.42645,5.07543-6.91245,12.31749-5.26594,18.95377q12.98639-3.06819,25.97253-6.13643l-.94771.66968A20.91064,20.91064,0,0,1,755.89151,256.134a21.66552,21.66552,0,0,1-4.57058,17.99384q8.85671-3.14187,17.71341-6.28364c3.64451-1.29283,7.54507-2.78248,9.70615-6.09858C781.24617,257.90093,780.664,252.805,779.96517,248.21676Z" transform="translate(-272.47643 -213.29715)" fill="#2f2e41" /><path id="fd990326-c176-439f-9263-07fe4201e145-633" data-name="Path 2960" d="M273.72343,686.47368h652.5532a1.247,1.247,0,0,0,0-2.494H273.72343a1.247,1.247,0,0,0,0,2.494Z" transform="translate(-272.47643 -213.29715)" fill="#ccc" /></svg>
-    );
-};
-
-const SaveTimeSVG = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" data-name="Layer 1" width="571.75671" height="700.46347" viewBox="0 0 571.75671 700.46347" xmlnsXlink="http://www.w3.org/1999/xlink"><polygon points="237.145 687.3 249.405 687.3 255.239 640.012 237.145 640.012 237.145 687.3" fill="#ffb8b8" /><path d="M548.13974,783.06523h24.1438a15.40422,15.40422,0,0,1,15.387,15.38672v.5H548.13974Z" transform="translate(-314.12165 -99.76826)" fill="#2f2e41" /><polygon points="79.05 664.069 89.559 670.383 118.914 632.854 103.404 623.535 79.05 664.069" fill="#ffb8b8" /><path d="M384.3707,772.41338l8.18188-13.61768,20.6958,12.43408a15.40476,15.40476,0,0,1,5.26465,21.11377l-.25757.42871Z" transform="translate(-314.12165 -99.76826)" fill="#2f2e41" /><path d="M547.93124,777.15849c-2.47314-2.51806-8.6167-43.875-17.70678-106.68164-3.16114-21.84131-6.15284-42.51269-8.61768-58.34326-21.09351,46.37012-55.19556,96.53076-96.1748,141.43506l-.23609.25879-.32373-.1333c-13.10962-5.39844-15.54956-6.64942-21.44433-9.67188-1.7605-.90283-3.83106-1.96435-6.58887-3.34326l-.15259-.11816c-1.26611-1.44629,10.28491-20.61963,50.67212-85.813,17.59351-28.3999,35.78613-57.76611,47.8352-78.15332l.11231-.18994.21606-.04492c29.12769-6.07471,57.21314,1.05566,79.08277,20.07714l.17749.15479-.00586.23486c-.59009,23.90772-.09448,61.44824.343,94.57031.854,64.667.79809,83.55909-1.21631,84.15772l-.10132.019c-2.95239.24219-5.40283.48438-7.66187.70752a164.62926,164.62926,0,0,1-17.84936,1.02735l-.21143.001Z" transform="translate(-314.12165 -99.76826)" fill="#2f2e41" /><path d="M484.43067,396.30226c-4.12706-17.0404-7.62358-36.26017.15974-51.95461,5.40044-10.88957,25.37863-14.5567,36.831-10.63361s19.73248,14.5667,23.37408,26.17159,3.21841,24.11,1.63126,36.17411C525.36428,395.65513,504.30181,395.25052,484.43067,396.30226Z" transform="translate(-314.12165 -99.76826)" fill="#2f2e41" /><circle cx="205.85253" cy="271.96835" r="24.56103" fill="#ffb8b8" /><path d="M489.94857,366.67718a30.84951,30.84951,0,1,1,61.66193.81287C530.5881,367.53468,509.56571,367.57931,489.94857,366.67718Z" transform="translate(-314.12165 -99.76826)" fill="#2f2e41" /><path d="M586.45224,606.97734c-4.529-15.45117-6.78491-26.26514-9.39014-38.75391-3.57153-17.12011-7.61963-36.52441-18.77075-74.30322l-.02759-.09326.00928-.09619c3.15747-32.76465-10.6062-67.5332-35.91943-90.73828a8.7642,8.7642,0,0,0-11.96387.606c-14.241,15.99219-23.19092,30.97168-27.3606,45.79541l-.09008.32129c5.94531,19.13037,12.59887,38.356,19.03613,56.95606l.09228.26611-.18017.2168c-24.1001,29.01709-37.4292,67.80469-36.62622,106.52051a214.65391,214.65391,0,0,1,71.77685-12.01709A232.68628,232.68628,0,0,1,586.45224,606.97734Z" transform="translate(-314.12165 -99.76826)" fill="#3f3d56" /><polygon points="186.941 382.463 223.217 461.44 243.941 440.463 186.941 382.463" opacity="0.1" /><path d="M698.7525,180.275c.345-44.18146-28.476-80.2248-64.37336-80.5051s-65.27769,35.30857-65.62268,79.49c-.31277,40.05573,23.34783,73.42112,54.52046,79.51237l-2.15206,12.89936,23.82252.186L642.997,258.92628C674.261,253.32257,698.43972,220.33073,698.7525,180.275Z" transform="translate(-314.12165 -99.76826)" fill="#6c63ff" /><path d="M590.59182,200.43107l-8.99973-.07027a146.49361,146.49361,0,0,1,.39823-50.99845l8.99972.07028Q584.94116,175.83637,590.59182,200.43107Z" transform="translate(-314.12165 -99.76826)" fill="#fff" /><rect x="619.6142" y="258.76497" width="27" height="6" transform="translate(-312.05841 -104.70382) rotate(0.44739)" fill="#3f3d56" /><path d="M823.82828,204.66981c3.498-44.04411-22.67651-82.05289-58.46226-84.895s-67.63166,30.55863-71.12967,74.60274C691.065,234.30874,712.28343,269.278,742.9417,277.579l-3.06739,12.71282,23.74847,1.88612-1.02227-13.03763C794.18471,275.78268,820.65691,244.601,823.82828,204.66981Z" transform="translate(-314.12165 -99.76826)" fill="#ff6584" /><path d="M714.50469,217.05338l-8.97175-.71254a146.4936,146.4936,0,0,1,4.03774-50.83991l8.97175.71254Q710.62414,192.11805,714.50469,217.05338Z" transform="translate(-314.12165 -99.76826)" fill="#fff" /><rect x="739.04026" y="278.26627" width="27" height="6" transform="translate(-289.4912 -158.46502) rotate(4.54094)" fill="#3f3d56" /><rect x="466.79411" y="439.17102" width="376.13849" height="3.0002" transform="translate(-377.12196 669.09304) rotate(-58.50618)" fill="#3f3d56" /><path d="M874.62251,247.9391c20.78829-38.98675,11.96138-84.2842-19.71528-101.17464s-74.208,1.02208-94.99628,40.00883c-18.847,35.34612-13.351,75.87841,11.44542,95.72721l-7.887,10.4317,21.02155,11.209,4.26692-12.362C819.057,301.30811,855.77547,283.28522,874.62251,247.9391Z" transform="translate(-314.12165 -99.76826)" fill="#6c63ff" /><path d="M769.443,215.65476l-7.94157-4.23456A146.4937,146.4937,0,0,1,785.49729,166.418l7.94157,4.23456Q775.83846,191.24312,769.443,215.65476Z" transform="translate(-314.12165 -99.76826)" fill="#fff" /><rect x="765.18522" y="286.71272" width="27" height="6" transform="translate(-86.23403 -432.07333) rotate(28.06716)" fill="#3f3d56" /><rect x="479.38779" y="441.3691" width="376.13796" height="3.0001" transform="translate(-397.45648 622.28469) rotate(-53.96275)" fill="#3f3d56" /><rect x="415.30797" y="445.48174" width="376.13841" height="2.9998" transform="matrix(0.17486, -0.98459, 0.98459, 0.17486, -256.3467, 863.1347)" fill="#3f3d56" /><path d="M581.93882,566.17943a10.05577,10.05577,0,0,1-9.055-12.48052L542.9932,534.11506l17.04692-7.36451,25.77121,19.72622a10.11027,10.11027,0,0,1-3.87251,19.70266Z" transform="translate(-314.12165 -99.76826)" fill="#ffb8b8" /><path d="M552.87485,551.72343c-17.875-20.564-36.35865-41.82812-53.18579-63.206l-.10791-.14355.00073-.17432c.0769-17.457.156-35.5044,2.22387-55.69531a15.98269,15.98269,0,0,1-4.23657-13.08741,17.72123,17.72123,0,0,1,7.78174-13.23388c8.42236-5.32032,21.04175-2.43994,27.00659,6.1665,5.759,8.30811,4.91748,20.21.5232,26.20459l-4.865,38.15234-.43018-.0332.09521.04834.0337-.02783.063.07666.14869.07568c-.00806.01612-.01807.03858-.03028.06934,15.78443,19.27148,32.10327,39.19287,48.40454,58.26514l.36133.42285-.45874.31445c-7.63745,5.23633-15.53491,10.65088-22.64868,15.88672l-.377.2666Z" transform="translate(-314.12165 -99.76826)" fill="#3f3d56" /><path d="M696.12165,800.23174h-381a1,1,0,0,1,0-2h381a1,1,0,0,1,0,2Z" transform="translate(-314.12165 -99.76826)" fill="#3f3d56" /><polygon points="492.715 480.274 492.715 480.274 492.715 480.274 492.715 480.274" fill="#f2f2f2" /><path d="M846.12842,591.45957l-.6051,1.90672-37.67848-11.95547c3.34052,4.41256,14.533,18.1937,26.16243,21.23118,13.89316,3.62856,21.048-1.44333,23.14159-9.45878s-1.66672-15.9385-15.56012-19.56747c-11.77683-3.07615-28.553,3.67819-33.39278,5.808Z" transform="translate(-314.12165 -99.76826)" fill="#f2f2f2" /><polygon points="350.832 578.731 350.832 578.73 350.832 578.73 350.832 578.731" fill="#f2f2f2" /><path d="M668.02152,719.30068l-1.99788.10107-1.99886-39.47918c-2.963,4.67443-11.951,19.98482-10.72281,31.94147,1.46742,14.284,8.72405,19.20922,16.965,18.36271s14.34511-7.14366,12.878-21.428c-1.24358-12.10826-13.446-25.456-17.136-29.24323Z" transform="translate(-314.12165 -99.76826)" fill="#f2f2f2" /><polygon points="63.981 131.667 63.981 131.667 63.981 131.667 63.981 131.667" fill="#f2f2f2" /><path d="M344.65347,207.87l1.19183-1.60663,31.74905,23.54992c-1.72543-5.25857-7.83252-21.92876-17.84388-28.58027-11.96018-7.94608-20.37486-5.47432-24.95933,1.42589s-3.60277,15.61518,8.3575,23.56174c10.1382,6.736,28.19878,5.7994,33.468,5.35779Z" transform="translate(-314.12165 -99.76826)" fill="#f2f2f2" /></svg>
-);
-
-export const LoginImageSlider: FC = () => {
-    const FIRST_SLIDE = "first";
-    const SECOND_SLIDE = "second";
-    const [currentSlide, setCurrentSlide] = useState(FIRST_SLIDE);
-
-    useEffect(() => {
-        setTimeout(() => {
-            const newSlide = currentSlide === FIRST_SLIDE ? SECOND_SLIDE : FIRST_SLIDE;
-            setCurrentSlide(newSlide);
-        }, 10000);
-    }, [currentSlide]);
-
-    // Render
-    return (
-      <div className={styles.loginImageSlider}>
-        <div className={[styles.itemContainer, currentSlide === FIRST_SLIDE ? styles.currentSlide : ""].join(" ")}>
-            <div className={styles.textContainer}>
-                <div className={styles.title}>
-                    Capture Information
-                </div>
-                <p className={styles.subTitle}>
-                    Capture information for R&D projects, tasks or activities
-                </p>
-            </div>
-            <CollectInformationSVG />
-        </div>
-        <div className={[styles.itemContainer, currentSlide === SECOND_SLIDE ? styles.currentSlide : ""].join(" ")}>
-            <div className={styles.textContainer}>
-                <div className={styles.title}>
-                    Save Time
-                </div>
-                <p className={styles.subTitle}>
-                    Build technology knowledge center for later use
-                </p>
-            </div>
-            <SaveTimeSVG />
-        </div>
-      </div>
-    );
-};
-
-export default LoginImageSlider;
